requirements.txt
alpaca-trade-api==0.42
appdirs==1.4.3
art==4.4
asyncio-nats-client==0.10.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coverage==5.0.3
cssselect==1.1.0
Deprecated==1.2.7
fake-useragent==0.1.11
idna==2.8
lxml==4.4.2
multitasking==0.0.9
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
parse==1.14.0
psutil==5.6.6
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
scapy==2.4.3
six==1.13.0
soupsieve==1.9.5
tqdm==4.41.1
tweepy==3.8.0
urllib3==1.24.3
virtualenv==16.7.9
w3lib==1.21.0
websocket-client==0.57.0
websockets==8.1
wrapt==1.11.2
yahoo-fin==0.8.4
yfinance==0.1.54
alpha_vantage==2.2.0
